LayoutTests/imported/w3c/web-platform-tests/dom/nodes/MutationObserver-callback-arguments-expected.txt


PASS Callback is invoked with |this| value of MutationObserver and two arguments
